Help disambiguate success or failure of a Promise
Success and failure are considered too application-domain specific and
lead to ambiguities when reading code wrote against that API.

Consider the CAF (Auth Framework) use case where the result of an
authentication request is itself a success or a failure.

This change replace respectively `Success` with `Result` and `Failure` with
`Exception`: they better describe what the outcome of a `Promise` could be.